+++ src/al/artofsoul/BatBatGame/GameState/Level1State.java
@@ -218,8 +218,8 @@ public class Level1State extends GameState {
 
         // update tilemap
         tileMap.setPosition(
-                GamePanel.WIDTH / 2 - player.getx(),
-                GamePanel.HEIGHT / 2 - player.gety()
+                (double)GamePanel.WIDTH / 2 - player.getx(),
+                (double)GamePanel.HEIGHT / 2 - player.gety()
         );
         tileMap.update();
         tileMap.fixBounds();
