+++ src/al/artofsoul/BatBatGame/GameState/Level2State.java
@@ -215,7 +215,7 @@ public class Level2State extends GameState {
         player.update();
 
         // update tilemap
-        tileMap.setPosition(GamePanel.WIDTH / 2 - player.getx(), GamePanel.HEIGHT / 2 - player.gety());
+        tileMap.setPosition((double)GamePanel.WIDTH / 2 - player.getx(), (double)GamePanel.HEIGHT / 2 - player.gety());
         tileMap.update();
         tileMap.fixBounds();
