+++ src/al/artofsoul/BatBatGame/GameState/Level3State.java
@@ -219,8 +219,8 @@ public class Level3State extends GameState {
 
         // update tilemap
         tileMap.setPosition(
-                GamePanel.WIDTH / 2 - player.getx(),
-                GamePanel.HEIGHT / 2 - player.gety()
+                (double)GamePanel.WIDTH / 2 - player.getx(),
+                (double)GamePanel.HEIGHT / 2 - player.gety()
         );
         tileMap.update();
         tileMap.fixBounds();
